Update tarefa 4 avaliativa.cpp

+++ tarefa 4 avaliativa.cpp
@@ -5,16 +5,16 @@ quociente da divisão do primeiro pelo segundo número.
 2. (\) ? m \ i: quociente da divisão de m por i
 Exemplo: 5 mod 2 é 1
  5 \ 2 é 2*/
- #include<stdio.h>
+#include<stdio.h>
  int main(){
- 	double A, B, RESULTADO;
+ 	int A, B, RESULTADO;
 	printf("Digite o numero A:\n");
- 	scanf("%lf", &A);
+ 	scanf("%d", &A);
  	printf("Digite o numero B:\n");
- 	scanf("%lf", &B);
- 	RESULTADO = A%B;
- 	printf("Numero de A:%lf\n", A);
- 	printf("Numero de B:%lf\n", B);
- 	printf("resultado da divisao:%lf\n", RESULTADO);
+ 	scanf("%d", &B);
+ 	RESULTADO = A % B;
+ 	printf("Numero de A:%d\n", A);
+ 	printf("Numero de B:%d\n", B);
+ 	printf("resultado da divisao:%d\n", RESULTADO);
  	return 0;
  }
